fix: Address code review feedback
- Add debounce timer cleanup on component unmount
- Add AbortController to cancel pending fetch requests
- Sanitize search query to escape single quotes (security)
- Extract loadScript utility to reduce code duplication
- Handle AbortError gracefully in async operations
